Read me once
I just want to tell you that the smart phone are not as smart as they have to be cause these smart device were only used for the entertainment purpose and not for the "health and security purpose" .
&#9734;First of all these smart devices has to be work in the off mode too in an emergency situation (cause in an emergency we don't have time to grab the phone then switch it on and to call or message the needy person )
So my suggestion is that we have to make a coded device that device is ready for all the emergency situation. Situation like 111 is for heart attack , 222 is for kidnapping , 333 is for accident , 444 is for robbery and 555 is for hijacking and so on (for example:- we code 666 for blast . So as soon as someone call 666 even the phone were switched off the phone scan the nearby fire station along with the hospital and start alerting them off (in the police station too if needed) at same time with real time location now the needy person start reacting) so that how the smart device save people from a big trouble cause it inform all the important nearby department which is directly or indirectly related with incident.
If it is not possible for the device to scan the nearby police station or the hospital and alerting them of so just build a "call centre" ( these call centre only work for the code or for the emergency situation). The smart device activated code alert the call centre along with real time location now its call centre agent turn to take action and scan the nearby needy department and inform them about the incident.
That how you generate job vacancy too.
Please read it out once.
Thank you .
 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 
 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 
 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 
 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 	 
 	 	 	 	I just want to tell you that the smart phone are not as smart as they have to be cause these smart device were only used for the entertainment purpose and not for the "health and security purpose" .
&#9734;First of all these smart devices has to be work in the off mode too in an emergency situation (cause in an emergency we don't have time to grab the phone then switch it on and to call or message the needy person )
So my suggestion is that we have to make a coded device that device is ready for all the emergency situation. Situation like 111 is for heart attack , 222 is for kidnapping , 333 is for accident , 444 is for robbery and 555 is for hijacking and so on (for example:- we code 666 for blast . So as soon as someone call 666 even the phone were switched off the phone scan the nearby fire station along with the hospital and start alerting them off (in the police station too if needed) at same time with real time location now the needy person start reacting) so that how the smart device save people from a big trouble cause it inform all the important nearby department which is directly or indirectly related with incident.
If it is not possible for the device to scan the nearby police station or the hospital and alerting them of so just build a "call centre" ( these call centre only work for the code or for the emergency situation). The smart device activated code alert the call centre along with real time location now its call centre agent turn to take action and scan the nearby needy department and inform them about the incident.
That how you generate job vacancy too.
Please read it out once.
Thank you.